* commit 'faef28c5610084213af9058aa9cab508c1d5e087':
  Core/Player: Remove unnecessary cast * 02131fe32bb6837e75bdeae0ca94c4ee06ac26bf followup

+++ src/server/game/Entities/Player/Player.cpp
@@ -2318,7 +2318,7 @@ void Player::GiveLevel(uint8 level)
     packet.PowerDelta[5] = 0;
 
     for (uint8 i = STAT_STRENGTH; i < MAX_STATS; ++i)
-        packet.StatDelta[i] = int32(info.stats[i]) - GetCreateStat(Stats(i));
+        packet.StatDelta[i] = info.stats[i] - GetCreateStat(Stats(i));
 
     uint32 const* rowLevels = (GetClass() != CLASS_DEATH_KNIGHT) ? DefaultTalentRowLevels : DKTalentRowLevels;
